fps_control

+++ src/Engine/Param_engine.h
@@ -14,7 +14,7 @@ struct Param_engine{
 
   std::string title = "Nephos";
 
-  int max_fps = 60;
+  float max_fps = 60.0f;
 
   glm::vec4 background_color = glm::vec4(0.01, 0.01, 0.01, 0.01);
   glm::vec2 window_dim = glm::vec2(1024, 500);
